[Cleanup] Moving a bunch of classes into android package (pt3). (#204)
* Move BackCallback into android package.

* Move FragmentLifecycleEvent into android.events package.

* Move Integration into android package.

* Move remaining fragments classes into android package. (#205)
