+++ formula-android-tests/src/main/java/com/instacart/formula/test/TestFragmentComponent.kt
@@ -3,9 +3,9 @@ package com.instacart.formula.test
 import android.view.View
 import com.instacart.formula.RenderView
 import com.instacart.formula.Renderer
-import com.instacart.formula.fragment.FragmentComponent
-import com.instacart.formula.fragment.FragmentContract
-import com.instacart.formula.fragment.FragmentLifecycleCallback
+import com.instacart.formula.android.FragmentComponent
+import com.instacart.formula.android.FragmentContract
+import com.instacart.formula.android.FragmentLifecycleCallback
 
 object TestFragmentComponent {
