+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentFlowRenderViewTest.kt
@@ -6,10 +6,10 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
-import com.instacart.formula.fragment.FragmentContract
-import com.instacart.formula.fragment.FragmentFlowState
+import com.instacart.formula.android.FragmentContract
+import com.instacart.formula.android.FragmentFlowState
 import com.instacart.formula.android.FragmentKey
-import com.instacart.formula.integration.BackCallback
+import com.instacart.formula.android.BackCallback
 import com.instacart.formula.test.TestContract
 import com.instacart.formula.test.TestContractWithId
 import com.instacart.formula.test.TestFragmentActivity
