+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentLifecycleStateTest.kt
@@ -4,8 +4,8 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
-import com.instacart.formula.fragment.FormulaFragment
-import com.instacart.formula.fragment.FragmentContract
+import com.instacart.formula.android.FormulaFragment
+import com.instacart.formula.android.FragmentContract
 import com.instacart.formula.android.ActivityStoreContext
 import com.instacart.formula.test.TestContract
 import com.instacart.formula.test.TestContractWithId
