+++ formula-android/src/test/java/com/instacart/formula/android/fakes/FakeAuthFlowFactory.kt
@@ -2,7 +2,7 @@ package com.instacart.formula.android.fakes
 
 import com.instacart.formula.android.Flow
 import com.instacart.formula.android.FlowFactory
-import com.instacart.formula.fragment.FragmentContract
+import com.instacart.formula.android.FragmentContract
 import com.instacart.formula.android.DisposableScope
 import io.reactivex.rxjava3.core.Observable
