+++ samples/todoapp/src/main/java/com/examples/todoapp/TodoActivity.kt
@@ -3,7 +3,7 @@ package com.examples.todoapp
 import android.os.Bundle
 import android.widget.Toast
 import com.examples.todoapp.tasks.TaskListKey
-import com.instacart.formula.fragment.FormulaFragment
+import com.instacart.formula.android.FormulaFragment
 import com.instacart.formula.android.FormulaAppCompatActivity
 
 class TodoActivity : FormulaAppCompatActivity() {
